pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "square-survivor"
version = "1.0.0"
authors = [{ name = "AntiGravity", email = "antigravity@gemini.com" }]
description = "A 2D Top-Down Roguelike Survival Game built with pygame-ce."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pygame-ce>=2.5.0"
]

[project.scripts]
square-survivor = "square_survivor.main:main"

[tool.setuptools.packages.find]
where = ["src"]
